requirements.txt
numpy==2.1.3
python-dotenv==1.0.1
Requests==2.32.3
boto3==1.28.66
mysql-connector-python==8.0.33
